+++ Services/RequestServer/main.cpp
@@ -16,6 +16,7 @@
 #include <LibIPC/SingleServer.h>
 #include <LibMain/Main.h>
 #include <LibTLS/Certificate.h>
+#include <LibTLS/TLSv12.h>
 #include <RequestServer/ConnectionFromClient.h>
 
 #if defined(AK_OS_MACOS)
